Move sampler options to options_models
