+++ qiskit_ibm_runtime/executor/routines/sampler_v2/__init__.py
@@ -13,4 +13,4 @@
 """Executor-based SamplerV2"""
 
 from .sampler import SamplerV2
-from ..options import SamplerOptions
+from ....options_models import SamplerOptions
