+++ qiskit_ibm_runtime/options_models/__init__.py
@@ -32,6 +32,7 @@ Base primitive options
 
    ExecutorOptions
    NoiseLearnerV3Options
+   SamplerOptions
 
 
 Suboptions
@@ -52,3 +53,4 @@ from .execution_options import ExecutionOptions
 from .executor_options import ExecutorOptions
 from .noise_learner_v3_options import NoiseLearnerV3Options
 from .post_selection_options import PostSelectionOptions
+from .sampler_options import SamplerOptions
