+++ test/unit/executor/routines/sampler_v2/test_sampler_v2_options.py
@@ -14,7 +14,7 @@
 
 import unittest
 
-from qiskit_ibm_runtime.executor.routines.options import SamplerOptions
+from qiskit_ibm_runtime.options_models import SamplerOptions
 
 
 class TestSamplerOptionsToExecutorOptions(unittest.TestCase):
