+++ test/unit/executor/routines/sampler_v2/test_sampler_v2_post_processor.py
@@ -23,7 +23,7 @@ from qiskit_ibm_runtime.executor.routines.sampler_v2 import SamplerV2
 from qiskit_ibm_runtime.executor.routines.sampler_v2.post_processors.v0_1 import (
     sampler_v2_post_processor_v0_1,
 )
-from qiskit_ibm_runtime.executor.routines.options.sampler_options import SamplerOptions
+from qiskit_ibm_runtime.options_models.sampler_options import SamplerOptions
 from qiskit_ibm_runtime.quantum_program.quantum_program_result import (
     QuantumProgramResult,
     Metadata,
